完成初诊部分数据连接

+++ hospital-back-end/src/main/java/com/neuedu/hospitalbackend/service/serviceimplementation/doctorstationservice/DiagnoseTemplateServiceImpl.java
@@ -108,7 +108,7 @@ public class DiagnoseTemplateServiceImpl implements DiagnoseTemplateService {
 
         }
         //若该诊断已存该疾病，但更新内容中无该疾病，则删除该疾病
-        while(!existedDiseaseIcdCodes.isEmpty()) {
+        if(!existedDiseaseIcdCodes.isEmpty()) {
             for (String leftDiseaseIcdCode : existedDiseaseIcdCodes)
                 count = diagnoseTemplateMapper.deleteByDiseaseIcdCode(leftDiseaseIcdCode);
         }
